python-backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
anthropic==0.52.0
python-dotenv==1.1.0
